# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/masters/{id}/services": {
      "get": {
        "operationId": "services",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MasterServiceResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "A master’s active services",
        "tags": [
          "Masters"
        ]
      }
    },
...
    "/messages/{id}": {
      "delete": {
        "description": "Sender-side soft delete.",
        "operationId": "remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "UNAUTHORIZED | TOKEN_EXPIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN | ACCOUNT_BLOCKED | ACCOUNT_INACTIVE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete the caller’s own message",
        "tags": [
          "Chat"
        ]
      }
    },
...
    "/admin/banners": {
      "get": {
        "description": "Includes inactive banners.",
        "operationId": "list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedDto"
                    },
                    {
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/BannerResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED | TOKEN_EXPIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — requires role: ADMIN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List every banner",
        "tags": [
          "Admin"
        ]
      },
      "post": {
        "operationId": "create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBannerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BannerResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED | TOKEN_EXPIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — requires role: ADMIN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "FILE_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a banner",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/banners/{id}": {
      "get": {
        "operationId": "findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BannerResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED | TOKEN_EXPIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — requires role: ADMIN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "BANNER_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Fetch a single banner",
        "tags": [
          "Admin"
        ]
      },
      "patch": {
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBannerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BannerResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED | TOKEN_EXPIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — requires role: ADMIN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "BANNER_NOT_FOUND | FILE_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a banner",
        "tags": [
          "Admin"
        ]
      },
      "delete": {
        "description": "Soft delete.",
        "operationId": "remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "UNAUTHORIZED | TOKEN_EXPIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — requires role: ADMIN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "BANNER_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a banner",
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "UstoGo API",
    "description": "Marketplace API connecting clients with professional craftsmen.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    },
    {
      "url": "https://api.staging.ustogo.app",
      "description": "Staging"
    },
    {
      "url": "https://api.ustogo.app",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token from /auth/login"
      }
    },
    "schemas": {
...
      "ErrorDetailDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "example": "email",
            "description": "Dot path; array elements use indices"
          },
          "constraints": {
            "example": [
              "email must be an email"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "field",
          "constraints"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 409,
            "description": "Mirrors the HTTP response status"
          },
          "code": {
            "type": "string",
            "example": "SLOT_NOT_AVAILABLE",
            "description": "Stable identifier. Clients branch on this, never on `message`."
          },
          "message": {
            "type": "string",
            "example": "The requested time slot is no longer available.",
            "description": "English, developer-facing. Clients localise from `code`."
          },
          "details": {
            "description": "Populated for 422; empty otherwise",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDetailDto"
            }
          },
          "path": {
            "type": "string",
            "example": "/api/v1/bookings",
            "description": "Path without the query string"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-07-29T10:15:00.000Z",
            "description": "ISO-8601 UTC"
          },
          "requestId": {
            "type": "string",
            "example": "3f1a2b8c-9d4e-4f60-8a1b-2c3d4e5f6a7b",
            "description": "Correlates to logs and to the X-Request-Id header"
          }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "details",
          "path",
          "timestamp",
          "requestId"
        ]
      },
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 137
          },
          "totalPages": {
            "type": "number",
            "example": 7
          },
          "hasNext": {
            "type": "boolean",
            "example": true
          },
          "hasPrev": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages",
          "hasNext",
          "hasPrev"
        ]
      },
      "PaginatedDto": {
        "type": "object",
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "meta"
        ]
      },
...
      "MasterServiceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "priceType": {
            "type": "string",
            "enum": [
              "FIXED",
              "HOURLY",
              "FROM"
            ]
          },
          "price": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "durationMinutes": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "categoryId",
          "title",
          "priceType",
          "price",
          "currency",
          "durationMinutes"
        ]
      },
...
      "BannerResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "subtitle": {
            "type": "object",
            "nullable": true
          },
          "imageFileId": {
            "type": "string",
            "format": "uuid"
          },
          "linkUrl": {
            "type": "object",
            "nullable": true
          },
          "position": {
            "type": "string",
            "enum": [
              "HOME_TOP",
              "HOME_MIDDLE",
              "CATEGORY_TOP"
            ]
          },
          "sortOrder": {
            "type": "number"
          },
          "startsAt": {
            "type": "object",
            "nullable": true
          },
          "endsAt": {
            "type": "object",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "createdByUserId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "imageFileId",
          "position",
          "sortOrder",
          "isActive",
          "createdByUserId",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateBannerDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200
          },
          "subtitle": {
            "type": "string",
            "maxLength": 500
          },
          "imageKey": {
            "type": "string",
            "format": "uuid",
            "description": "A confirmed File with purpose 'BANNER'."
          },
          "linkUrl": {
            "type": "string",
            "maxLength": 500
          },
          "position": {
            "type": "string",
            "enum": [
              "HOME_TOP",
              "HOME_MIDDLE",
              "CATEGORY_TOP"
            ]
          },
          "sortOrder": {
            "type": "number",
            "default": 0
          },
          "startsAt": {
            "type": "string",
            "description": "ISO-8601. Omit for \"always started\"."
          },
          "endsAt": {
            "type": "string",
            "description": "ISO-8601. Omit for \"never expires\"."
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "title",
          "imageKey",
          "position"
        ]
      },
      "UpdateBannerDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200
          },
          "subtitle": {
            "type": "string",
            "maxLength": 500
          },
          "imageKey": {
            "type": "string",
            "format": "uuid",
            "description": "A confirmed File with purpose 'BANNER'."
          },
          "linkUrl": {
            "type": "string",
            "maxLength": 500
          },
          "position": {
            "type": "string",
            "enum": [
              "HOME_TOP",
              "HOME_MIDDLE",
              "CATEGORY_TOP"
            ]
          },
          "sortOrder": {
            "type": "number"
          },
          "startsAt": {
            "type": "string",
            "description": "ISO-8601."
          },
          "endsAt": {
            "type": "string",
            "description": "ISO-8601."
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}